service/requirements.txt
streamlit
sqlalchemy
pydantic
celery
redis
passlib
pyjwt==2.1.0